kafka-pipeline-haasa/consumer/requirements.txt
confluent_kafka==1.3.0
pymongo==3.10.1
python-dotenv==0.12.0
tweepy==3.8.0
dnspython==1.16.0
